+++ tests/testthat/test-tar_make_clustermq.R
@@ -33,3 +33,71 @@ tar_test("tar_make_clustermq() can use tidyselect", {
   out <- sort(list.files(file.path("_targets", "objects")))
   expect_equal(out, sort(c("y1", "y2")))
 })
+
+tar_test("custom script and store args", {
+  skip_on_cran()
+  expect_equal(tar_config_get("script"), path_script_default())
+  expect_equal(tar_config_get("store"), path_store_default())
+  old_option <- getOption("clustermq.scheduler")
+  on.exit(options(clustermq.scheduler = old_option))
+  tar_script({
+    options(clustermq.scheduler = "multicore")
+    tar_target(x, getNamespace("targets")$tar_config$is_locked())
+  }, script = "example/script.R")
+  tar_make_clustermq(
+    script = "example/script.R",
+    store = "example/store",
+    callr_function = NULL
+  )
+  expect_false(file.exists("_targets.yaml"))
+  expect_equal(tar_config_get("script"), path_script_default())
+  expect_equal(tar_config_get("store"), path_store_default())
+  expect_equal(path_script(), path_script_default())
+  expect_equal(path_store(), path_store_default())
+  expect_false(file.exists(path_script_default()))
+  expect_false(file.exists(path_store_default()))
+  expect_true(file.exists("example/script.R"))
+  expect_true(file.exists("example/store"))
+  expect_true(file.exists("example/store/meta/meta"))
+  expect_true(file.exists("example/store/objects/x"))
+  expect_equal(readRDS("example/store/objects/x"), TRUE)
+  expect_false(tar_config$is_locked())
+  tar_config_set(script = "x")
+  expect_equal(tar_config_get("script"), "x")
+  expect_true(file.exists("_targets.yaml"))
+})
+
+tar_test("custom script and store args with callr function", {
+  skip_on_cran()
+  expect_equal(tar_config_get("script"), path_script_default())
+  expect_equal(tar_config_get("store"), path_store_default())
+  tar_script({
+    options(clustermq.scheduler = "multicore")
+    tar_target(x, getNamespace("targets")$tar_config$is_locked())
+  }, script = "example/script.R")
+  tmp <- utils::capture.output(
+    suppressMessages(
+      tar_make_clustermq(
+        script = "example/script.R",
+        store = "example/store",
+        reporter = "silent"
+      )
+    )
+  )
+  expect_false(file.exists("_targets.yaml"))
+  expect_equal(tar_config_get("script"), path_script_default())
+  expect_equal(tar_config_get("store"), path_store_default())
+  expect_equal(path_script(), path_script_default())
+  expect_equal(path_store(), path_store_default())
+  expect_false(file.exists(path_script_default()))
+  expect_false(file.exists(path_store_default()))
+  expect_true(file.exists("example/script.R"))
+  expect_true(file.exists("example/store"))
+  expect_true(file.exists("example/store/meta/meta"))
+  expect_true(file.exists("example/store/objects/x"))
+  expect_equal(readRDS("example/store/objects/x"), TRUE)
+  expect_false(tar_config$is_locked())
+  tar_config_set(script = "x")
+  expect_equal(tar_config_get("script"), "x")
+  expect_true(file.exists("_targets.yaml"))
+})